+++ services/users/fixtures/TestListUsers.golden
@@ -4,6 +4,7 @@
    "name": "AdminO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     }
    ]
@@ -12,6 +13,7 @@
    "name": "AdminO2",
    "orgs": [
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
@@ -19,13 +21,17 @@
   {
    "name": "OrgAdmin",
    "orgs": [
-    {}
+    {
+     "name": "\u003croot\u003e",
+     "id": "root"
+    }
    ]
   },
   {
    "name": "UserO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     }
    ]
@@ -34,6 +40,7 @@
    "name": "UserO2",
    "orgs": [
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
@@ -44,6 +51,7 @@
    "name": "AdminO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     }
    ]
@@ -52,6 +60,7 @@
    "name": "UserO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     }
    ]
@@ -62,6 +71,7 @@
    "name": "AdminO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     }
    ]
@@ -70,6 +80,7 @@
    "name": "UserO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     }
    ]
@@ -77,10 +88,22 @@
  ],
  "UserO1 ListUsers": [
   {
-   "name": "AdminO1"
+   "name": "AdminO1",
+   "orgs": [
+    {
+     "name": "O1",
+     "id": "O1"
+    }
+   ]
   },
   {
-   "name": "UserO1"
+   "name": "UserO1",
+   "orgs": [
+    {
+     "name": "O1",
+     "id": "O1"
+    }
+   ]
   }
  ],
  "AdminO2 ListUsers - Filtered AdminO1 Orgs": [
@@ -88,6 +111,7 @@
    "name": "AdminO1",
    "orgs": [
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
@@ -96,6 +120,7 @@
    "name": "AdminO2",
    "orgs": [
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
@@ -104,6 +129,7 @@
    "name": "UserO2",
    "orgs": [
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
@@ -114,9 +140,11 @@
    "name": "AdminO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     },
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
@@ -125,6 +153,7 @@
    "name": "AdminO2",
    "orgs": [
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
@@ -133,6 +162,7 @@
    "name": "UserO1",
    "orgs": [
     {
+     "name": "O1",
      "id": "O1"
     }
    ]
@@ -141,6 +171,7 @@
    "name": "UserO2",
    "orgs": [
     {
+     "name": "O2",
      "id": "O2"
     }
    ]
